// src/enums/ErrorCode.ts
export enum ErrorCode {
  EMAIL_EXISTS = "EMAIL_EXISTS",
  INVALID_ROLE = "INVALID_ROLE",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  INVALID_TOKEN = "TOKEN_INVALID",
}